Refactor(lumenfall): Simplify structure and fix paths
This commit refactors the LumenFall game from a single monolithic HTML file into a simple and clean structure with separate files for HTML, CSS, and JavaScript.

- **New Directory Structure:** The game now resides in `lumenfall/` with `css/` and `js/` subdirectories.
- **Separated Files:** The code is now in `lumenfall/index.html`, `lumenfall/css/style.css`, and a single `lumenfall/js/game.js`.
- **Applied Fixes:** This version includes the fixes for audio paths and mobile joystick controls that were developed previously.
- **Updated Asset Paths:** All asset paths within the CSS and JS have been updated to be relative to the new file structure.
- **Updated Portfolio Link:** The main `index.html` now correctly links to the new game location.

This simplified structure resolves the loading issues experienced with the more complex module-based refactoring and provides a clean, maintainable foundation for future work.
